+++ src/Business/Grand.Business.Catalog/Queries/Models/GetPriceByCustomerProductQuery.cs
@@ -2,7 +2,7 @@
 
 namespace Grand.Business.Catalog.Queries.Models
 {
-    public class GetPriceByCustomerProductQuery : IRequest<decimal?>
+    public class GetPriceByCustomerProductQuery : IRequest<double?>
     {
         public string CustomerId { get; set; }
         public string ProductId { get; set; }
